updated atoi to return long long, added itoa_base

+++ ft_atoi.c
@@ -6,14 +6,14 @@
 /*   By: kchiang <kchiang@student.42kl.edu.my>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/20 15:28:26 by kchiang           #+#    #+#             */
-/*   Updated: 2025/06/08 18:04:11 by kchiang          ###   ########.fr       */
+/*   Updated: 2025/06/08 18:43:23 by kchiang          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 /* converts string digits to integer */
 /* only handles one sign character */
 /* the moment it hits a non digit, return nbr */
-int	ft_atoi(const char *nptr)
+long long	ft_atoi(const char *nptr)
 {
 	int			sign;
 	long long	nbr;
